Refactored save method to module
