+++ lib/foods.js
@@ -2,6 +2,7 @@ var foodsTable = document.getElementById('foods-table');
 var trashIcon = document.getElementById('trash-icon')
 var foodForm = document.getElementById('food-form')
 var Reset = require('./reset')
+var Save = require('./save')
 
 function Food(name, calories){
   this.name = name;
@@ -83,6 +84,7 @@ Food.prototype.editValue = function(obj, key){
   var name = this.name;
   var calories = this.calories
   var type = 'foods';
+  var save = new Save;
   obj.contentEditable = true;
   clickedEntry = obj;
   var originalKey = obj.innerText;
@@ -91,24 +93,15 @@ Food.prototype.editValue = function(obj, key){
     if (event.keyCode == 13) {
       var newKey = obj.innerText;
       obj.contentEditable = false;
-      setLocalStorage(key, newKey, name, calories, type);
+      save.setLocalStorage(key, newKey, name, calories, type);
     }
   });
   $(document).click(function(e){
     var newKey = clickedEntry.innerText;
     clickedEntry.contentEditable = false;
-    setLocalStorage(key, newKey, name, calories, type);
+    save.setLocalStorage(key, newKey, name, calories, type);
   });
 }
-
-// function emptyErrors(idName){
-//   $(idName).children().children().empty();
-// }
-//
-// function clearFields(field1, field2){
-//   $(field1).val("");
-//   $(field2).val("");
-// }
 
 Food.prototype.storeFoods = function() {
   var foodDataJSON = localStorage.getItem('foods');
@@ -130,17 +123,6 @@ Food.prototype.displayFoods = function(){
       <td><a href='' id='trash-icon'><span class='glyphicon glyphicon-trash trash-icon'></a></td>
       </tr>`);
   })
-}
-
-function setLocalStorage(key, newKey, name, calories, type){
-  var array = JSON.parse(localStorage.getItem(type))
-  array.forEach(function(object) {
-    if (object.name === name && object.calories === calories) {
-      object[key] = newKey;
-    }
-    arrayJSON = JSON.stringify(array);
-    localStorage.setItem(type, arrayJSON);
-    })
 }
 
 function filterItems(table, input) {
